% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{gbm_all_predictions}
\alias{gbm_all_predictions}
\title{Compute all iterative boosting predictions}
\usage{
gbm_all_predictions(gbm, X_star)
}
\arguments{
\item{gbm}{A gradient boosting model of class "qc_basement_gbm"}

\item{X_star}{The data to predict for (as a data frame). It has n_* rows and p columns}
}
\value{
A matrix with n_* rows and M+1 columns where each column are the iterative
               predictions across all base learners beginning with g_0. For regression, the
               unit is in the units of the original response. For probability estimation for 
               binary response, the unit is the logit of the probability estimate.
}
\description{
Returns all predictions for each iteration of the gradient boosting
}
\author{
Adam Kapelner
}
