% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{predict.qc_basement_gbm}
\alias{predict.qc_basement_gbm}
\title{GBM Predict}
\usage{
\method{predict}{qc_basement_gbm}(gbm, X_star)
}
\arguments{
\item{gbm}{A gradient boosting model of class "qc_basement_gbm"}

\item{X_star}{The data to predict for (as a data frame). It has n_* rows and p columns}
}
\value{
A vector of length n_* rows with each row's predictions. For regression, the
               unit is in the units of the original response. For probability estimation for 
               binary response, the unit is the logit of the probability estimate.
}
\description{
Returns final predictions for the gradient boosting model
}
\author{
Adam Kapelner
}
