% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{print.qc_basement_gbm}
\alias{print.qc_basement_gbm}
\title{Prints a summary of a \code{qc_basement_gbm} object}
\usage{
\method{print}{qc_basement_gbm}(x, ...)
}
\arguments{
\item{...}{Other parameters to pass to the default print function}

\item{object}{The \code{qc_basement_gbm} object to be summarized in the console}
}
\description{
Prints a summary of a \code{qc_basement_gbm} object
}
\author{
Adam Kapelner
}
