% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCbasementGBM.R
\docType{package}
\name{QCbasementGBM}
\alias{QCbasementGBM}
\title{QC Basement Gradient Boosting Model}
\description{
Creates fits, inference and predictions for shifted and scaled beta models of continuous phenomena
}
\references{
Kapelner, A
}
\author{
Adam Kapelner \email{kapelner@qc.cuny.edu}
}
\keyword{inference,}
\keyword{modeling,}
\keyword{prediction}
\keyword{statistical}
