% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{summary.qc_basement_gbm}
\alias{summary.qc_basement_gbm}
\title{Prints a summary of a \code{qc_basement_gbm} object}
\usage{
\method{summary}{qc_basement_gbm}(object, ...)
}
\arguments{
\item{object}{The \code{qc_basement_gbm} object to be summarized in the console}

\item{...}{Other parameters to pass to the default summary function}
}
\description{
Prints a summary of a \code{qc_basement_gbm} object
}
\author{
Adam Kapelner
}
